+++ BetterContinents.WorldGeneratorPatch.cs
@@ -69,52 +69,25 @@ public partial class BetterContinents
 
         // wx, wy are [-10500, 10500]
         // __result should be [0, 1]
-        [HarmonyPrefix, HarmonyPatch(nameof(WorldGenerator.GetBaseHeight))]
-        private static bool GetBaseHeightPrefix(ref float wx, ref float wy, bool menuTerrain, ref float __result, float ___m_offset0, float ___m_offset1, float ___m_minMountainDistance)
+        public static bool GetBaseHeightPrefixV3(ref float wx, ref float wy, ref float __result, float ___m_minMountainDistance)
         {
-            if (!Settings.EnabledForThisWorld || menuTerrain)
-            {
-                return true;
-            }
-
-            switch (Settings.Version)
-            {
-                case 1:
-                case 2:
-                    __result = GetBaseHeightV1(wx, wy, ___m_offset0, ___m_offset1, ___m_minMountainDistance);
-                    break;
-                case 3:
-                case 4:
-                case 5:
-                case 6:
-                    __result = GetBaseHeightV2(wx, wy, ___m_offset0, ___m_offset1, ___m_minMountainDistance);
-                    break;
-                case 7:
-                default:
-                    // GetBaseHeightV3 doesn't work at all without heightmap which makes testing more difficult.
-                    if (Settings.HasHeightmap || BaseHeightNoise != null)
-                        __result = GetBaseHeightV3(wx, wy, ___m_minMountainDistance);
-                    else return true;
-                    break;
-            }
-
+            __result = GetBaseHeightV3(wx, wy, ___m_minMountainDistance);
+            return false;
+        }
+        public static bool GetBaseHeightPrefixV2(ref float wx, ref float wy, ref float __result, float ___m_offset0, float ___m_offset1, float ___m_minMountainDistance)
+        {
+            __result = GetBaseHeightV2(wx, wy, ___m_offset0, ___m_offset1, ___m_minMountainDistance);
+            return false;
+        }
+        public static bool GetBaseHeightPrefixV1(ref float wx, ref float wy, ref float __result, float ___m_offset0, float ___m_offset1, float ___m_minMountainDistance)
+        {
+            __result = GetBaseHeightV1(wx, wy, ___m_offset0, ___m_offset1, ___m_minMountainDistance);
             return false;
         }
 
-        [HarmonyPostfix, HarmonyPatch(nameof(WorldGenerator.GetBiomeHeight))]
-        private static void GetBiomeHeightPostfix(WorldGenerator __instance, float wx, float wy, ref float __result, World ___m_world)
+        public static void GetBiomeHeightPostfix(WorldGenerator __instance, float wx, float wy, ref float __result)
         {
-            if (!Settings.EnabledForThisWorld || ___m_world.m_menu)
-            {
-                return;
-            }
-
-            if (!Settings.ShouldHeightmapOverrideAll && !Settings.HasRoughmap)
-            {
-                return;
-            }
-
-            float smoothHeight = __instance.GetBaseHeight(wx, wy, false) * 200f;
+            var smoothHeight = __instance.GetBaseHeight(wx, wy, false) * 200f;
             __result = Settings.ShouldHeightmapOverrideAll
                 ? smoothHeight
                 : Settings.ApplyRoughmap(NormalizedX(wx), NormalizedY(wy), smoothHeight, __result);
@@ -307,278 +280,32 @@ public partial class BetterContinents
             return finalHeight;
         }
 
-
-
-
-        // We must come after WorldGenOptions, as that mod always replaces the GetBiome function.
-        // Note in HarmonyX (which BepInEx uses), all prefixes always run, unless they opt out themselves.
-        // So coming BEFORE a prefex and returning false doesn't stop that prefix from running.
-        [HarmonyPrefix, HarmonyPatch(nameof(WorldGenerator.GetBiome), typeof(float), typeof(float)), HarmonyAfter("expand_world_data")]
-        private static bool GetBiomePrefix(float wx, float wy, ref Heightmap.Biome __result, World ___m_world)
+        public static bool GetBiomePrefix(float wx, float wy, ref Heightmap.Biome __result, World ___m_world)
         {
-            if (!Settings.EnabledForThisWorld || ___m_world.m_menu || !Settings.HasBiomemap)
-            {
-                return true;
-            }
-            else
-            {
-                var normalized = WorldToNormalized(wx, wy);
-                var result = Settings.GetBiomeOverride(normalized.x, normalized.y);
-                if (result != Heightmap.Biome.None)
-                    __result = result;
-                return false;
-            }
+            var normalized = WorldToNormalized(wx, wy);
+            var result = Settings.GetBiomeOverride(normalized.x, normalized.y);
+            if (result != Heightmap.Biome.None)
+                __result = result;
+            return false;
         }
 
-        [HarmonyPrefix, HarmonyPatch(nameof(WorldGenerator.AddRivers))]
-        private static bool AddRiversPrefix(ref float __result, float h, World ___m_world)
+        public static bool AddRiversPrefix(ref float __result, float h)
         {
-            if (!Settings.EnabledForThisWorld || ___m_world.m_menu || Settings.RiversEnabled)
-            {
-                // Fall through to normal function
-                return true;
-            }
-            else
-            {
-                __result = h;
-                return false;
-            }
+            __result = h;
+            return false;
         }
 
-        [HarmonyPrefix, HarmonyPatch(nameof(WorldGenerator.GetForestFactor))]
-        private static void GetForestFactorPrefix(ref Vector3 pos, ref object __state)
+        public static void GetForestFactorPrefix(ref Vector3 pos)
         {
-            // Save the pos before modifying it, for use in the postfix below (to save possible float accuracy
-            // weirdness if we were to attempt to reverse the scaling instead)
-            __state = pos;
-            if (Settings.EnabledForThisWorld && Settings.ForestScale != 1)
-            {
-                pos *= Settings.ForestScale;
-            }
+            pos *= Settings.ForestScale;
         }
 
         // Range: 0.145071 1.850145
-        [HarmonyPostfix, HarmonyPatch(nameof(WorldGenerator.GetForestFactor))]
-        private static void GetForestFactorPostfix(object __state, ref float __result)
+        public static void GetForestFactorPostfix(Vector3 pos, ref float __result)
         {
-            if (Settings.EnabledForThisWorld)
-            {
-                // We stored the original unmodified pos in the prefix above
-                var pos = (Vector3)__state;
-                __result = Settings.ApplyForest(NormalizedX(pos.x), NormalizedY(pos.z), __result);
-            }
-        }
-    }
-}
-
-// None of this is faster in testing
-[HarmonyPatch(typeof(HeightmapBuilder))]
-public class HeightmapBuilderPatch
-{
-    [HarmonyPrefix, HarmonyPatch(nameof(HeightmapBuilder.Build))]
-    private static bool BuildPrefix(HeightmapBuilder.HMBuildData data, out Stopwatch __state)
-    {
-        __state = new Stopwatch();
-        __state.Start();
-        if (BetterContinents.ConfigExperimentalMultithreadedHeightmapBuild.Value)
-        {
-            BuildMT(data);
-            return false;
-        }
-        else
-        {
-            return true;
-        }
-    }
-
-    private static int patchBuildCount = 0;
-    [HarmonyPostfix, HarmonyPatch(nameof(HeightmapBuilder.Build))]
-    private static void BuildPostfix(Stopwatch __state)
-    {
-        if (patchBuildCount++ % 20 == 0)
-        {
-            BetterContinents.Log(BetterContinents.ConfigExperimentalMultithreadedHeightmapBuild.Value
-                ? $"Heightmap Build MT {__state.ElapsedMilliseconds} ms"
-                : $"Heightmap Build Vanilla {__state.ElapsedMilliseconds} ms"
-            );
-        }
-    }
-
-    private static List<float> buildTimes = [];
-    [HarmonyPrefix, HarmonyPatch(nameof(HeightmapBuilder.BuildThread))]
-    private static bool BuildThreadPrefix(HeightmapBuilder __instance)
-    {
-        ZLog.Log("Builder started");
-        while (!__instance.m_stop)
-        {
-            if (!BetterContinents.ConfigExperimentalParallelChunksBuild.Value)
-            {
-                __instance.m_lock.WaitOne();
-                bool flag = __instance.m_toBuild.Count > 0;
-                __instance.m_lock.ReleaseMutex();
-                if (flag)
-                {
-                    __instance.m_lock.WaitOne();
-                    HeightmapBuilder.HMBuildData hmbuildData = __instance.m_toBuild[0];
-                    __instance.m_lock.ReleaseMutex();
-
-                    var st = new Stopwatch();
-                    st.Start();
-                    __instance.Build(hmbuildData);
-                    buildTimes.Add(st.ElapsedMilliseconds);
-                    if (buildTimes.Count > 100)
-                    {
-                        buildTimes.RemoveAt(0);
-                    }
-                    // BetterContinents.Log($"Average chunk build time with vanilla {buildTimes.Average()} ms");
-
-                    __instance.m_lock.WaitOne();
-                    __instance.m_toBuild.Remove(hmbuildData);
-                    __instance.m_ready.Add(hmbuildData);
-                    while (__instance.m_ready.Count > 16)
-                    {
-                        __instance.m_ready.RemoveAt(0);
-                    }
-                    __instance.m_lock.ReleaseMutex();
-                }
-                Thread.Sleep(10);
-            }
-            else
-            {
-                __instance.m_lock.WaitOne();
-                var buildTasks = __instance.m_toBuild
-                    .Select(b => new
-                    {
-                        task = Task.Run(() => __instance.Build(b)),
-                        buildData = b
-                    })
-                    .ToList(); // Important to use ToList for force immediate enumeration inside the lock!
-                __instance.m_toBuild.Clear();
-                __instance.m_lock.ReleaseMutex();
-
-                if (buildTasks.Any())
-                {
-                    var st = new Stopwatch();
-                    st.Start();
-                    Task.WaitAll(buildTasks.Select(bt => bt.task).ToArray());
-                    buildTimes.Add(st.ElapsedMilliseconds / (float)buildTasks.Count);
-                    if (buildTimes.Count > 100)
-                    {
-                        buildTimes.RemoveAt(0);
-                    }
-                    // BetterContinents.Log($"Average chunk build time with MT {buildTimes.Average()} ms");
-
-                    __instance.m_lock.WaitOne();
-                    foreach (var bt in buildTasks)
-                    {
-                        __instance.m_ready.Add(bt.buildData);
-                    }
-
-                    while (__instance.m_ready.Count > 16)
-                    {
-                        __instance.m_ready.RemoveAt(0);
-                    }
-
-                    __instance.m_lock.ReleaseMutex();
-                }
-            }
-
-            Thread.Sleep(10);
-        }
-
-        return false;
-    }
-
-    // This is slower, at least for the 64x64 tiles
-    private static void BuildMT(HeightmapBuilder.HMBuildData data)
-    {
-        int num = data.m_width + 1;
-        int num2 = num * num;
-        Vector3 vector = data.m_center + new Vector3(data.m_width * data.m_scale * -0.5f, 0f, data.m_width * data.m_scale * -0.5f);
-        WorldGenerator worldGen = data.m_worldGen;
-        data.m_cornerBiomes = new Heightmap.Biome[4];
-        data.m_cornerBiomes[0] = worldGen.GetBiome(vector.x, vector.z);
-        data.m_cornerBiomes[1] = worldGen.GetBiome(vector.x + data.m_width * data.m_scale, vector.z);
-        data.m_cornerBiomes[2] = worldGen.GetBiome(vector.x, vector.z + data.m_width * data.m_scale);
-        data.m_cornerBiomes[3] = worldGen.GetBiome(vector.x + data.m_width * data.m_scale, vector.z + data.m_width * data.m_scale);
-        Heightmap.Biome biome = data.m_cornerBiomes[0];
-        Heightmap.Biome biome2 = data.m_cornerBiomes[1];
-        Heightmap.Biome biome3 = data.m_cornerBiomes[2];
-        Heightmap.Biome biome4 = data.m_cornerBiomes[3];
-        data.m_baseHeights = new List<float>(num * num);
-        for (int i = 0; i < num2; i++)
-        {
-            data.m_baseHeights.Add(0f);
-        }
-
-        GameUtils.SimpleParallelFor(4, 0, num, j =>
-        //for (int j = 0; j < num; j++)
-        {
-            float wy = vector.z + j * data.m_scale;
-            float t = Mathf.SmoothStep(0f, 1f, j / (float)data.m_width);
-            for (int k = 0; k < num; k++)
-            {
-                float wx = vector.x + k * data.m_scale;
-                float t2 = Mathf.SmoothStep(0f, 1f, k / (float)data.m_width);
-                float value;
-                if (data.m_distantLod)
-                {
-                    Heightmap.Biome biome5 = worldGen.GetBiome(wx, wy);
-                    value = worldGen.GetBiomeHeight(biome5, wx, wy, out _);
-                }
-                else if (biome3 == biome && biome2 == biome && biome4 == biome)
-                {
-                    value = worldGen.GetBiomeHeight(biome, wx, wy, out _);
-                }
-                else
-                {
-                    float biomeHeight = worldGen.GetBiomeHeight(biome, wx, wy, out _);
-                    float biomeHeight2 = worldGen.GetBiomeHeight(biome2, wx, wy, out _);
-                    float biomeHeight3 = worldGen.GetBiomeHeight(biome3, wx, wy, out _);
-                    float biomeHeight4 = worldGen.GetBiomeHeight(biome4, wx, wy, out _);
-                    float a = Mathf.Lerp(biomeHeight, biomeHeight2, t2);
-                    float b = Mathf.Lerp(biomeHeight3, biomeHeight4, t2);
-                    value = Mathf.Lerp(a, b, t);
-                }
-
-                data.m_baseHeights[j * num + k] = value;
-            }
-        });
-
-        if (data.m_distantLod)
-        {
-            for (int l = 0; l < 4; l++)
-            {
-                List<float> list = new(data.m_baseHeights);
-                for (int m = 1; m < num - 1; m++)
-                {
-                    for (int n = 1; n < num - 1; n++)
-                    {
-                        float num3 = list[m * num + n];
-                        float num4 = list[(m - 1) * num + n];
-                        float num5 = list[(m + 1) * num + n];
-                        float num6 = list[m * num + n - 1];
-                        float num7 = list[m * num + n + 1];
-                        if (Mathf.Abs(num3 - num4) > 10f)
-                        {
-                            num3 = (num3 + num4) * 0.5f;
-                        }
-                        if (Mathf.Abs(num3 - num5) > 10f)
-                        {
-                            num3 = (num3 + num5) * 0.5f;
-                        }
-                        if (Mathf.Abs(num3 - num6) > 10f)
-                        {
-                            num3 = (num3 + num6) * 0.5f;
-                        }
-                        if (Mathf.Abs(num3 - num7) > 10f)
-                        {
-                            num3 = (num3 + num7) * 0.5f;
-                        }
-                        data.m_baseHeights[m * num + n] = num3;
-                    }
-                }
-            }
+            if (Settings.ForestScale != 1f)
+                pos /= Settings.ForestScale;
+            __result = Settings.ApplyForest(NormalizedX(pos.x), NormalizedY(pos.z), __result);
         }
     }
 }